Rearrange the etcd/felix support code
