+++ fv/containers/containers.go
@@ -16,7 +16,6 @@ package containers
 
 import (
 	"bufio"
-	"context"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -31,9 +30,6 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/projectcalico/felix/fv/utils"
-	api "github.com/projectcalico/libcalico-go/lib/apis/v3"
-	client "github.com/projectcalico/libcalico-go/lib/clientv3"
-	"github.com/projectcalico/libcalico-go/lib/options"
 	"github.com/projectcalico/libcalico-go/lib/set"
 )
 
@@ -424,200 +420,4 @@ func (c *Container) CanConnectTo(ip, port, protocol string) bool {
 		"stderr": string(wErr)}).WithError(err).Info("Connection test")
 
 	return err == nil
-}
-
-func RunEtcd() *Container {
-	log.Info("Starting etcd")
-	return Run("etcd",
-		RunOpts{AutoRemove: true},
-		"--privileged", // So that we can add routes inside the etcd container,
-		// when using the etcd container to model an external client connecting
-		// into the cluster.
-		utils.Config.EtcdImage,
-		"etcd",
-		"--advertise-client-urls", "http://127.0.0.1:2379",
-		"--listen-client-urls", "http://0.0.0.0:2379")
-}
-
-type Felix struct {
-	*Container
-}
-
-func (f *Felix) GetFelixPID() int {
-	return f.GetSinglePID("calico-felix")
-}
-
-func (f *Felix) GetFelixPIDs() []int {
-	return f.GetPIDs("calico-felix")
-}
-
-func RunFelix(etcdIP string, options TopologyOptions) *Felix {
-	log.Info("Starting felix")
-	ipv6Enabled := fmt.Sprint(options.EnableIPv6)
-
-	args := []string{
-		"--privileged",
-		"-e", "CALICO_DATASTORE_TYPE=etcdv3",
-		"-e", "CALICO_ETCD_ENDPOINTS=http://" + etcdIP + ":2379",
-		"-e", "FELIX_LOGSEVERITYSCREEN=" + options.FelixLogSeverity,
-		"-e", "FELIX_DATASTORETYPE=etcdv3",
-		"-e", "FELIX_PROMETHEUSMETRICSENABLED=true",
-		"-e", "FELIX_USAGEREPORTINGENABLED=false",
-		"-e", "FELIX_IPV6SUPPORT=" + ipv6Enabled,
-		"-v", "/lib/modules:/lib/modules",
-	}
-
-	for k, v := range options.ExtraEnvVars {
-		args = append(args, "-e", fmt.Sprintf("%s=%s", k, v))
-	}
-
-	for k, v := range options.ExtraVolumes {
-		args = append(args, "-v", fmt.Sprintf("%s:%s", k, v))
-	}
-
-	args = append(args,
-		utils.Config.FelixImage,
-	)
-
-	c := Run("felix",
-		RunOpts{AutoRemove: true},
-		args...,
-	)
-
-	if options.EnableIPv6 {
-		c.Exec("sysctl", "-w", "net.ipv6.conf.all.disable_ipv6=0")
-		c.Exec("sysctl", "-w", "net.ipv6.conf.default.disable_ipv6=0")
-		c.Exec("sysctl", "-w", "net.ipv6.conf.lo.disable_ipv6=0")
-		c.Exec("sysctl", "-w", "net.ipv6.conf.all.forwarding=1")
-	} else {
-		c.Exec("sysctl", "-w", "net.ipv6.conf.all.disable_ipv6=1")
-		c.Exec("sysctl", "-w", "net.ipv6.conf.default.disable_ipv6=1")
-		c.Exec("sysctl", "-w", "net.ipv6.conf.lo.disable_ipv6=1")
-		c.Exec("sysctl", "-w", "net.ipv6.conf.all.forwarding=0")
-	}
-
-	return &Felix{
-		Container: c,
-	}
-}
-
-type TopologyOptions struct {
-	FelixLogSeverity      string
-	EnableIPv6            bool
-	ExtraEnvVars          map[string]string
-	ExtraVolumes          map[string]string
-	AlphaFeaturesToEnable string
-}
-
-func DefaultTopologyOptions() TopologyOptions {
-	return TopologyOptions{
-		FelixLogSeverity: "info",
-		EnableIPv6:       true,
-		ExtraEnvVars:     map[string]string{},
-		ExtraVolumes:     map[string]string{},
-	}
-}
-
-// StartSingleNodeEtcdTopology starts an etcd container and a single Felix container; it initialises
-// the datastore and installs a Node resource for the Felix node.
-func StartSingleNodeEtcdTopology(options TopologyOptions) (felix *Felix, etcd *Container, calicoClient client.Interface) {
-	felixes, etcd, calicoClient := StartNNodeEtcdTopology(1, options)
-	felix = felixes[0]
-	return
-}
-
-// StartNNodeEtcdTopology starts an etcd container and a set of Felix hosts.  If n > 1, sets
-// up IPIP, otherwise this is skipped.
-//
-// - Configures an IPAM pool for 10.65.0.0/16 (so that Felix programs the all-IPAM blocks IP set)
-//   but (for simplicity) we don't actually use IPAM to assign IPs.
-// - Configures routes between the hosts, giving each host 10.65.x.0/24, where x is the
-//   index in the returned array.  When creating workloads, use IPs from the relevant block.
-// - Configures the Tunnel IP for each host as 10.65.x.1.
-func StartNNodeEtcdTopology(n int, opts TopologyOptions) (felixes []*Felix, etcd *Container, client client.Interface) {
-	log.Infof("Starting a %d-node etcd topology.", n)
-	success := false
-	var err error
-	defer func() {
-		if !success {
-			log.WithError(err).Error("Failed to start topology, tearing down containers")
-			for _, felix := range felixes {
-				felix.Stop()
-			}
-			etcd.Stop()
-		}
-	}()
-
-	// First start etcd.
-	etcd = RunEtcd()
-
-	// Connect to etcd.
-	client = utils.GetEtcdClient(etcd.IP, opts.AlphaFeaturesToEnable)
-	mustInitDatastore(client)
-
-	if n > 1 {
-		Eventually(func() error {
-			ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-			defer cancel()
-			ipPool := api.NewIPPool()
-			ipPool.Name = "test-pool"
-			ipPool.Spec.CIDR = "10.65.0.0/16"
-			ipPool.Spec.IPIPMode = api.IPIPModeAlways
-			_, err = client.IPPools().Create(ctx, ipPool, options.SetOptions{})
-			return err
-		}).ShouldNot(HaveOccurred())
-	}
-
-	for i := 0; i < n; i++ {
-		// Then start Felix and create a node for it.
-		felix := RunFelix(etcd.IP, opts)
-
-		felixNode := api.NewNode()
-		felixNode.Name = felix.Hostname
-		if n > 1 {
-			felixNode.Spec.BGP = &api.NodeBGPSpec{
-				IPv4Address:        felix.IP,
-				IPv4IPIPTunnelAddr: fmt.Sprintf("10.65.%d.1", i),
-			}
-		}
-		Eventually(func() error {
-			_, err = client.Nodes().Create(utils.Ctx, felixNode, utils.NoOptions)
-			if err != nil {
-				log.WithError(err).Warn("Failed to create node")
-			}
-			return err
-		}, "10s", "500ms").ShouldNot(HaveOccurred())
-
-		felixes = append(felixes, felix)
-	}
-
-	// Set up routes between the hosts, note: we're not using IPAM here but we set up similar
-	// CIDR-based routes.
-	for i, iFelix := range felixes {
-		for j, jFelix := range felixes {
-			if i == j {
-				continue
-			}
-
-			jBlock := fmt.Sprintf("10.65.%d.0/24", j)
-			err := iFelix.ExecMayFail("ip", "route", "add", jBlock, "via", jFelix.IP, "dev", "tunl0", "onlink")
-			Expect(err).ToNot(HaveOccurred())
-		}
-	}
-	success = true
-	return
-}
-
-func mustInitDatastore(client client.Interface) {
-	Eventually(func() error {
-		log.Info("Initializing the datastore...")
-		ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
-		err := client.EnsureInitialized(
-			ctx,
-			"v3.0.0-test",
-			"felix-fv",
-		)
-		log.WithError(err).Info("EnsureInitialized result")
-		return err
-	}).ShouldNot(HaveOccurred())
 }
