+++ fv/netsets_test.go
@@ -27,6 +27,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
+	"github.com/projectcalico/felix/fv/infrastructure"
 	"github.com/projectcalico/felix/fv/utils"
 	"github.com/projectcalico/libcalico-go/lib/options"
 
@@ -83,14 +84,14 @@ var _ = Context("Network sets tests with initialized Felix and etcd datastore", 
 
 	var (
 		etcd     *containers.Container
-		felix    *containers.Felix
+		felix    *infrastructure.Felix
 		felixPID int
 		client   client.Interface
 	)
 
 	BeforeEach(func() {
-		topologyOptions := containers.DefaultTopologyOptions()
-		felix, etcd, client = containers.StartSingleNodeEtcdTopology(topologyOptions)
+		topologyOptions := infrastructure.DefaultTopologyOptions()
+		felix, etcd, client = infrastructure.StartSingleNodeEtcdTopology(topologyOptions)
 		felixPID = felix.GetFelixPID()
 	})
